requirement.txt
ostorlab[agent]
rich
py-ubjson
